# Web APi/RavCevelGood1/bin/Release/net5.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "RavCevelGood",
    "version": "v1"
  },
  "paths": {
...
    "/api/Group/GetGroupsByIdSchool/{SchoolId}/{YearbookId}": {
      "get": {
        "tags": [
          "Group"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "SchoolId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "YearbookId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Group/GetStudentPerGroup/{GroupId}/{YearbookId}": {
      "get": {
        "tags": [
          "Group"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "GroupId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "YearbookId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Group/DeleteGroup/{GroupId}/{YearbookId}": {
      "delete": {
        "tags": [
          "Group"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "GroupId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "YearbookId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Group/AddGroup/{UserCreatedId}/{SchoolId}/{YearbookId}/{userId}": {
      "put": {
        "tags": [
          "Group"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "UserCreatedId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "SchoolId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "YearbookId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/AppGroupDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Group/GetAllNameGroup/{SchoolId}": {
      "get": {
        "tags": [
          "Group"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "SchoolId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/User/GetAllUserBySchoolId/{SchoolId}": {
      "get": {
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "SchoolId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/User/GetUsersBySchoolIDAndYearbookId/{SchoolId}/{YearbookId}": {
      "get": {
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "SchoolId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "YearbookId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/User/GetUserDetailsByIDUser/{UserId}/{SchoolId}": {
      "get": {
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "UserId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "SchoolId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/User/AddUser/{UserCreatedId}/{SchoolId}/{yearbookId}/{emailAddress}/{message}": {
      "put": {
        "tags": [
          "User"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "UserCreatedId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "SchoolId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "yearbookId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "emailAddress",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "name": "message",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/AppUserPerSchoolWithDetailsDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/User/DeleteUser/{UserID}/{SchoolId}": {
      "delete": {
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "UserID",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "SchoolId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "AppUserPerGroupDTO": {
      "type": "object",
      "properties": {
        "iduserPerGroup": {
          "format": "int32",
          "type": "integer"
        },
        "userId": {
          "format": "int32",
          "type": "integer"
        },
        "groupId": {
          "format": "int32",
          "type": "integer"
        },
        "fromDate": {
          "format": "date-time",
          "type": "string"
        },
        "toDate": {
          "format": "date-time",
          "type": "string"
        }
      }
    },
    "AppGroupPerYearbookDTO": {
      "type": "object",
      "properties": {
        "idgroupPerYearbook": {
          "format": "int32",
          "type": "integer"
        },
        "groupId": {
          "format": "int32",
          "type": "integer"
        },
        "yearbookId": {
          "format": "int32",
          "type": "integer"
        },
        "appUserPerGroups": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/AppUserPerGroupDTO"
          }
        }
      }
    },
    "AppGroupDTO": {
      "type": "object",
      "properties": {
        "idgroup": {
          "format": "int32",
          "type": "integer"
        },
        "ageGroupId": {
          "format": "int32",
          "type": "integer"
        },
        "schoolId": {
          "format": "int32",
          "type": "integer"
        },
        "typeGroupId": {
          "format": "int32",
          "type": "integer"
        },
        "extensionId": {
          "format": "int32",
          "type": "integer"
        },
        "voiceSpaceId": {
          "format": "int32",
          "type": "integer"
        },
        "listeningTimeId": {
          "format": "int32",
          "type": "integer"
        },
        "userCreated": {
          "format": "int32",
          "type": "integer"
        },
        "dateCreated": {
          "format": "date-time",
          "type": "string"
        },
        "userUpdated": {
          "format": "int32",
          "type": "integer"
        },
        "dateUpdated": {
          "format": "date-time",
          "type": "string"
        },
        "nameGroup": {
          "type": "string"
        },
        "idgroupPeryearbook": {
          "format": "int32",
          "type": "integer"
        },
        "appGroupPerYearbooks": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/AppGroupPerYearbookDTO"
          }
        }
      }
    },
    "AppAddressDTO": {
      "type": "object",
      "properties": {
        "idaddress": {
          "format": "int32",
          "type": "integer"
        },
        "cityId": {
          "format": "int32",
          "type": "integer"
        },
        "streetId": {
          "format": "int32",
          "type": "integer"
        },
        "houseNumber": {
          "type": "string"
        },
        "apartmentNumber": {
          "type": "string"
        },
        "poBox": {
          "format": "int32",
          "type": "integer"
        },
        "zipCode": {
          "format": "int32",
          "type": "integer"
        },
        "neighborhoodId": {
          "format": "int32",
          "type": "integer"
        },
        "userCreated": {
          "format": "int32",
          "type": "integer"
        },
        "dateCreated": {
          "format": "date-time",
          "type": "string"
        },
        "userUpdated": {
          "format": "int32",
          "type": "integer"
        },
        "dateUpdated": {
          "format": "date-time",
          "type": "string"
        },
        "comment": {
          "type": "string"
        }
      }
    },
    "AppBirthDTO": {
      "type": "object",
      "properties": {
        "idbirth": {
          "format": "int32",
          "type": "integer"
        },
        "birthCountryId": {
          "format": "int32",
          "type": "integer"
        },
        "birthDate": {
          "format": "date-time",
          "type": "string"
        },
        "birthHebrewDate": {
          "type": "string"
        },
        "citizenshipId": {
          "format": "int32",
          "type": "integer"
        },
        "dateOfImmigration": {
          "format": "date-time",
          "type": "string"
        },
        "countryIdofImmigration": {
          "format": "int32",
          "type": "integer"
        },
        "userCreated": {
          "format": "int32",
          "type": "integer"
        },
        "dateCreated": {
          "format": "date-time",
          "type": "string"
        },
        "userUpdated": {
          "format": "int32",
          "type": "integer"
        },
        "dateUpdated": {
          "format": "date-time",
          "type": "string"
        }
      }
    },
    "AppContactInformationDTO": {
      "type": "object",
      "properties": {
        "idcontactInformation": {
          "format": "int32",
          "type": "integer"
        },
        "telephoneNumber1": {
          "type": "string"
        },
        "telephoneNumber2": {
          "type": "string"
        },
        "phoneNumber1": {
          "type": "string"
        },
        "phoneNumber2": {
          "type": "string"
        },
        "phoneNumber3": {
          "type": "string"
        },
        "faxNumber": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "comment": {
          "type": "string"
        },
        "userCreated": {
          "format": "int32",
          "type": "integer"
        },
        "dateCreated": {
          "format": "date-time",
          "type": "string"
        },
        "userUpdated": {
          "format": "int32",
          "type": "integer"
        },
        "dateUpdated": {
          "format": "date-time",
          "type": "string"
        },
        "isMobile": {
          "type": "string"
        },
        "idadress": {
          "format": "int32",
          "type": "integer"
        },
        "job": {
          "type": "string"
        }
      }
    },
...
    "SecUserDTO": {
      "type": "object",
      "properties": {
        "iduser": {
          "format": "int32",
          "type": "integer"
        },
        "tz": {
          "type": "string"
        },
        "typeIdentityId": {
          "format": "int32",
          "type": "integer"
        },
        "userPassword": {
          "type": "string"
        },
        "userCode": {
          "type": "string"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "fullName": {
          "type": "string"
        },
        "genderId": {
          "format": "int32",
          "type": "integer"
        },
        "addressId": {
          "format": "int32",
          "type": "integer"
        },
        "userCreated": {
          "format": "int32",
          "type": "integer"
        },
        "dateCreated": {
          "format": "date-time",
          "type": "string"
        },
        "userUpdated": {
          "format": "int32",
          "type": "integer"
        },
        "dateUpdated": {
          "format": "date-time",
          "type": "string"
        },
        "birthId": {
          "format": "int32",
          "type": "integer"
        },
        "address": {
          "$ref": "#/definitions/AppAddressDTO"
        },
        "birth": {
          "$ref": "#/definitions/AppBirthDTO"
        }
      }
    },
    "AppUserPerSchoolWithDetailsDTO": {
      "type": "object",
      "properties": {
        "iduserPerSchool": {
          "format": "int32",
          "type": "integer"
        },
        "userId": {
          "format": "int32",
          "type": "integer"
        },
        "schoolId": {
          "format": "int32",
          "type": "integer"
        },
        "userCreated": {
          "format": "int32",
          "type": "integer"
        },
        "dateCreated": {
          "format": "date-time",
          "type": "string"
        },
        "userUpdeted": {
          "format": "int32",
          "type": "integer"
        },
        "dateUpdated": {
          "format": "date-time",
          "type": "string"
        },
        "contactInformationId": {
          "format": "int32",
          "type": "integer"
        },
        "statusId": {
          "format": "int32",
          "type": "integer"
        },
        "isActive": {
          "type": "boolean"
        },
        "note": {
          "type": "string"
        },
        "typeUserId": {
          "format": "int32",
          "type": "integer"
        },
        "user": {
          "$ref": "#/definitions/SecUserDTO"
        },
        "contactInformation": {
          "$ref": "#/definitions/AppContactInformationDTO"
        }
      }
    },
...
  }
}